Simplify ElectionSchedulerTests
