+++ server/src/test/java/org/elasticsearch/cluster/coordination/ElectionSchedulerTests.java
@@ -20,7 +20,6 @@
 package org.elasticsearch.cluster.coordination;
 
 import org.elasticsearch.Version;
-import org.elasticsearch.cluster.ClusterState.VotingConfiguration;
 import org.elasticsearch.cluster.coordination.CoordinationState.VoteCollection;
 import org.elasticsearch.cluster.node.DiscoveryNode;
 import org.elasticsearch.common.settings.Settings;
@@ -28,30 +27,19 @@ import org.elasticsearch.common.unit.TimeValue;
 import org.elasticsearch.test.ESTestCase;
 import org.elasticsearch.test.transport.CapturingTransport;
 import org.elasticsearch.threadpool.ThreadPool.Names;
-import org.elasticsearch.transport.ConnectTransportException;
 import org.elasticsearch.transport.Transport;
-import org.elasticsearch.transport.TransportRequest;
 import org.elasticsearch.transport.TransportService;
 import org.junit.Before;
 
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Objects;
-import java.util.stream.Collectors;
-
 import static java.util.Collections.emptySet;
-import static java.util.Collections.singleton;
+import static java.util.Collections.singletonList;
 import static org.elasticsearch.cluster.coordination.ElectionScheduler.ELECTION_BACK_OFF_TIME_SETTING;
 import static org.elasticsearch.cluster.coordination.ElectionScheduler.ELECTION_MAX_TIMEOUT_SETTING;
 import static org.elasticsearch.cluster.coordination.ElectionScheduler.ELECTION_MIN_TIMEOUT_SETTING;
 import static org.elasticsearch.cluster.coordination.ElectionScheduler.validationExceptionMessage;
 import static org.elasticsearch.cluster.coordination.PreVoteCollector.REQUEST_PRE_VOTE_ACTION_NAME;
 import static org.elasticsearch.node.Node.NODE_NAME_SETTING;
-import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.greaterThanOrEqualTo;
-import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.lessThanOrEqualTo;
 
@@ -61,46 +49,13 @@ public class ElectionSchedulerTests extends ESTestCase {
     private ElectionScheduler electionScheduler;
     private boolean electionOccurred = false;
     private DiscoveryNode localNode;
-    private Map<DiscoveryNode, PreVoteResponse> responsesByNode = new HashMap<>();
-    private VotingConfiguration votingConfiguration;
-    private long preVoteResponseDelay;
-    private long lastElectionMaxTermSeen;
 
     @Before
     public void createObjects() {
         final Settings settings = Settings.builder().put(NODE_NAME_SETTING.getKey(), "node").build();
         deterministicTaskQueue = new DeterministicTaskQueue(settings);
-        preVoteResponseDelay = 0L;
-        final Transport capturingTransport = new CapturingTransport() {
-            @Override
-            protected void onSendRequest(final long requestId, final String action, final TransportRequest request,
-                                         final DiscoveryNode node) {
-                super.onSendRequest(requestId, action, request, node);
-                assertThat(action, is(REQUEST_PRE_VOTE_ACTION_NAME));
-                assertThat(request, instanceOf(PreVoteRequest.class));
-                PreVoteRequest preVoteRequest = (PreVoteRequest) request;
-                assertThat(preVoteRequest.getSourceNode(), equalTo(localNode));
-                deterministicTaskQueue.scheduleAt(deterministicTaskQueue.getCurrentTimeMillis() + preVoteResponseDelay, new Runnable() {
-                    @Override
-                    public void run() {
-                        final PreVoteResponse response = responsesByNode.get(node);
-                        if (response == null) {
-                            handleRemoteError(requestId, new ConnectTransportException(node, "no response"));
-                        } else {
-                            handleResponse(requestId, response);
-                        }
-                    }
-
-                    @Override
-                    public String toString() {
-                        return "response to " + request + " from " + node;
-                    }
-                });
-            }
-        };
+        final Transport capturingTransport = new CapturingTransport();
         localNode = new DiscoveryNode("local-node", buildNewFakeTransportAddress(), Version.CURRENT);
-        responsesByNode.put(localNode, new PreVoteResponse(3, 2, 1));
-        votingConfiguration = new VotingConfiguration(singleton(localNode.getId()));
         final TransportService transportService = new TransportService(settings, capturingTransport,
             deterministicTaskQueue.getThreadPool(), TransportService.NOOP_TRANSPORT_INTERCEPTOR,
             boundTransportAddress -> localNode, null, emptySet());
@@ -108,33 +63,28 @@ public class ElectionSchedulerTests extends ESTestCase {
         transportService.acceptIncomingRequests();
 
         transportService.registerRequestHandler(REQUEST_PRE_VOTE_ACTION_NAME, Names.GENERIC, PreVoteRequest::new,
-            (request, channel, task) -> {
-                assertThat(request.getSourceNode(), equalTo(localNode));
-                // this is just used for local messages
-                channel.sendResponse(Objects.requireNonNull(responsesByNode.get(localNode)));
-            });
+            (request, channel, task) -> channel.sendResponse(new PreVoteResponse(3, 2, 1)));
 
         electionScheduler = new ElectionScheduler(settings, random(), transportService) {
             @Override
             protected void startElection(long maxTermSeen) {
                 assert electionOccurred == false;
                 electionOccurred = true;
-                lastElectionMaxTermSeen = maxTermSeen;
             }
 
             @Override
             protected Iterable<DiscoveryNode> getBroadcastNodes() {
-                return responsesByNode.keySet();
+                return singletonList(localNode);
             }
 
             @Override
             protected boolean isElectionQuorum(VoteCollection voteCollection) {
-                return votingConfiguration.hasQuorum(voteCollection.nodes().stream().map(DiscoveryNode::getId).collect(Collectors.toSet()));
+                return voteCollection.containsVoteFor(localNode);
             }
 
             @Override
             protected PreVoteResponse getLocalPreVoteResponse() {
-                return Objects.requireNonNull(responsesByNode.get(localNode));
+                return new PreVoteResponse(3, 2, 1);
             }
         };
     }
@@ -211,112 +161,6 @@ public class ElectionSchedulerTests extends ESTestCase {
             deterministicTaskQueue.runAllRunnableTasks(random());
         }
         electionOccurred = false;
-    }
-
-    public void testStartsElectionIfLocalNodeIsQuorum() {
-        final DiscoveryNode otherNode = new DiscoveryNode("other-node", buildNewFakeTransportAddress(), Version.CURRENT);
-        responsesByNode.put(otherNode, new PreVoteResponse(3, 2, 1));
-
-        electionScheduler.start(randomGracePeriod());
-        deterministicTaskQueue.advanceTime();
-        deterministicTaskQueue.runAllRunnableTasks(random());
-        assertTrue(electionOccurred);
-    }
-
-    public void testStartsElectionIfOtherNodeIsQuorum() {
-        final DiscoveryNode otherNode = new DiscoveryNode("other-node", buildNewFakeTransportAddress(), Version.CURRENT);
-        responsesByNode.put(otherNode, new PreVoteResponse(3, 2, 1));
-        votingConfiguration = new VotingConfiguration(singleton(otherNode.getId()));
-
-        electionScheduler.start(randomGracePeriod());
-        deterministicTaskQueue.advanceTime();
-        deterministicTaskQueue.runAllRunnableTasks(random());
-        assertTrue(electionOccurred);
-    }
-
-    public void testDoesNotStartsElectionIfOtherNodeIsQuorumAndDoesNotRespond() {
-        final DiscoveryNode otherNode = new DiscoveryNode("other-node", buildNewFakeTransportAddress(), Version.CURRENT);
-        responsesByNode.put(otherNode, null);
-        votingConfiguration = new VotingConfiguration(singleton(otherNode.getId()));
-
-        electionScheduler.start(randomGracePeriod());
-        deterministicTaskQueue.advanceTime();
-        deterministicTaskQueue.runAllRunnableTasks(random());
-
-        electionScheduler.stop();
-        deterministicTaskQueue.runAllTasks();
-
-        assertFalse(electionOccurred);
-    }
-
-    public void testDoesNotStartElectionIfStopped() {
-        final DiscoveryNode otherNode = new DiscoveryNode("other-node", buildNewFakeTransportAddress(), Version.CURRENT);
-        responsesByNode.put(otherNode, new PreVoteResponse(3, 2, 1));
-        votingConfiguration = new VotingConfiguration(singleton(otherNode.getId()));
-        preVoteResponseDelay = 1;
-
-        electionScheduler.start(randomGracePeriod());
-        deterministicTaskQueue.advanceTime();
-        deterministicTaskQueue.runAllRunnableTasks(random());
-
-        electionScheduler.stop();
-        deterministicTaskQueue.runAllTasks();
-        assertFalse(electionOccurred);
-    }
-
-    public void testIgnoresPreVotesFromLaterTerms() {
-        final DiscoveryNode otherNode = new DiscoveryNode("other-node", buildNewFakeTransportAddress(), Version.CURRENT);
-        responsesByNode.put(otherNode, new PreVoteResponse(3, 3, 1));
-        votingConfiguration = new VotingConfiguration(singleton(otherNode.getId()));
-
-        electionScheduler.start(randomGracePeriod());
-        deterministicTaskQueue.advanceTime();
-        deterministicTaskQueue.runAllRunnableTasks(random());
-
-        electionScheduler.stop();
-        deterministicTaskQueue.runAllTasks();
-        assertFalse(electionOccurred);
-    }
-
-    public void testIgnoresPreVotesFromLaterVersionInSameTerm() {
-        final DiscoveryNode otherNode = new DiscoveryNode("other-node", buildNewFakeTransportAddress(), Version.CURRENT);
-        responsesByNode.put(otherNode, new PreVoteResponse(3, 2, 2));
-        votingConfiguration = new VotingConfiguration(singleton(otherNode.getId()));
-
-        electionScheduler.start(randomGracePeriod());
-        deterministicTaskQueue.advanceTime();
-        deterministicTaskQueue.runAllRunnableTasks(random());
-
-        electionScheduler.stop();
-        deterministicTaskQueue.runAllTasks();
-        assertFalse(electionOccurred);
-    }
-
-    public void testAcceptsPreVotesFromLaterVersionInEarlierTerms() {
-        final DiscoveryNode otherNode = new DiscoveryNode("other-node", buildNewFakeTransportAddress(), Version.CURRENT);
-        responsesByNode.put(otherNode, new PreVoteResponse(3, 1, 2));
-        votingConfiguration = new VotingConfiguration(singleton(otherNode.getId()));
-
-        electionScheduler.start(randomGracePeriod());
-        deterministicTaskQueue.advanceTime();
-        deterministicTaskQueue.runAllRunnableTasks(random());
-        assertTrue(electionOccurred);
-        assertThat(lastElectionMaxTermSeen, is(3L));
-    }
-
-    public void testReturnsMaximumSeenTerm() {
-        final DiscoveryNode otherNode1 = new DiscoveryNode("other-node-1", buildNewFakeTransportAddress(), Version.CURRENT);
-        final DiscoveryNode otherNode2 = new DiscoveryNode("other-node-2", buildNewFakeTransportAddress(), Version.CURRENT);
-        responsesByNode.put(otherNode1, new PreVoteResponse(4, 2, 1));
-        responsesByNode.put(otherNode2, new PreVoteResponse(5, 2, 1));
-        votingConfiguration
-            = new VotingConfiguration(new HashSet<>(Arrays.asList(otherNode1.getId(), otherNode2.getId())));
-
-        electionScheduler.start(randomGracePeriod());
-        deterministicTaskQueue.advanceTime();
-        deterministicTaskQueue.runAllRunnableTasks(random());
-        assertTrue(electionOccurred);
-        assertThat(lastElectionMaxTermSeen, greaterThanOrEqualTo(5L));
     }
 
     public void testSettingsMustBeReasonable() {
